requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic[email]
pydantic-settings
pyjwt
python-multipart
slowapi
redis
alembic
